Re organize component
