+++ services/agora/src/components/post/analysis/opinionGroupTab/GroupConsensusSummary.vue
